Fix Hall of Fame page erroring for logged-out visitors (Task #592)
Problem
- /hall-of-fame rendered "Failed to load Hall of Fame data" because the
  /api/hall-of-fame endpoint 500'd. Root cause: getHallOfFameCareerStats in
  src/db/index.js used ROUND(AVG(...), 2) on a double precision value, which
  Postgres rejects ("function round(double precision, integer) does not exist").
  That call is the one parallel promise the frontend doesn't tolerate, so the
  whole page broke (not actually auth-specific; logged-in likely saw a cached
  response).

Changes
- src/db/index.js: cast AVG to ::numeric before ROUND in getHallOfFameCareerStats
  (avg_kda and avg_gpm), matching the sibling getPlayerBenchmarkAverages query.
- src/web/server.js: /hall-of-fame route now wraps getPersonalRecords,
  getHallOfFameCareerStats, and getImpactScores in individual .catch handlers so
  a single sub-query failure degrades gracefully instead of 500ing the page.
- src/data/patchNotes.js: added grouped entry v8.54.

Verification
- Reproduced the SQL throw against the live DB, confirmed fix returns 19 rows.
- Restarted workflow; GET /api/hall-of-fame returns 200; anon screenshot of
  /hall-of-fame renders records + all tabs correctly.
- npm run check:patch-notes passes (all versions unique).

Scope / deviations
- Full edition only, per task. The community edition (community-edition/src/db/index.js,
  line ~7060) has the identical ROUND bug but is a separate independent deploy and
  out of this task's scope — proposed as a follow-up.

+++ src/db/index.js
@@ -26689,8 +26689,8 @@ async function getHallOfFameCareerStats(seasonId = null) {
       COUNT(DISTINCT ps.match_id) AS games,
       SUM(CASE WHEN (ps.team = 'radiant' AND m.radiant_win) OR (ps.team = 'dire' AND NOT m.radiant_win) THEN 1 ELSE 0 END) AS wins,
       SUM(CASE WHEN (ps.team = 'radiant' AND NOT m.radiant_win) OR (ps.team = 'dire' AND m.radiant_win) THEN 1 ELSE 0 END) AS losses,
-      ROUND(AVG(CASE WHEN ps.deaths > 0 THEN (ps.kills + ps.assists)::float / ps.deaths ELSE (ps.kills + ps.assists)::float END), 2) AS avg_kda,
-      ROUND(AVG(ps.gpm)) AS avg_gpm,
+      ROUND(AVG(CASE WHEN ps.deaths > 0 THEN (ps.kills + ps.assists)::float / ps.deaths ELSE (ps.kills + ps.assists)::float END)::numeric, 2) AS avg_kda,
+      ROUND(AVG(ps.gpm)::numeric) AS avg_gpm,
       SUM(ps.kills) AS total_kills,
       (SELECT COUNT(*) FROM achievements a WHERE a.player_id = ps.account_id) AS achievement_count
     FROM player_stats ps
